Fix heading anchors scroll position
Currently when you scroll down to an anchor, by clicking an anchor link,
the heading you scrolled down to is hidden by the top header bar. That's
because it has a fixed positioned of being at the top.

This change uses a separate element for the anchor, which is positioned
above the heading element, so that the header doesn't move over it and
hide the heading you just scrolled down to.