Restore usage of index.html for overview pages
All the index.html were renamed to overview.html, to prevent matching on
them when there was a partial match on the file path.

Split the `path_active?` method into two methods that are more
descriptive in what they do: `section_active?` and `page_active?`, and
only use them where appropriate.

- `section_active?` will do a partial match on the file paths and is
  used to determine if a section should be opened or closed. If the
  page's path matches the section's path it will render as active.
  If the `section_prefix` is the root (`/`) it will only accept full
  matches, otherwise it will match on all sections as every section
  starts with a `/`.
- `page_active?` is only used for direct matches on filenames for
  documentation pages. This also works for pages using the `index.html`
  filename.

This way we don't have to jump through hoops, like querying the
Middleman Sitemap resource to find what paths actually are partial
filenames and if they are a directory index or not (usually a section).
Which I think will also be more resource intensive. And the current page
load speed locally already is not great.