Move anomaly detection page to features
It was in the applications section, but it's a bit hidden.

Anomaly detection is a feature on its own like metrics, so move it a
level higher and add it to the "Features" section in the navigation.

I used the "bell" icon that we also use on AppSignal.com for this
feature in the navigation.

This should also make it also makes it easier to find for people in the
docs.