Add parameter filtering guide
Merge the Ruby and Elixir parameter filtering pages into one guide.
I couldn't find any information about this "feature" in the JavaScript
or Node.js sections, and couldn't really tell if it's entirely
implemented based on their code bases, so I omitted these two
integrations for now. Feel free to add them if you are sure how they
work!

I added a new page that explains the parameter filtering in some more
detail, how the processor filtering works, etc.

I've moved the "filter out all parameters" part of the Ruby and Elixir
pages to the `/application/parameter-filtering.html` page because it
didn't feel like the best place for it was the guide. The "feature" page
seems like a better fit, for when people are reading further up on it.