Add request header collection guide
This guide is a bit different from the other filtering guides, as it
filters a lot by default, and is more of an "add more content" guide.

The feature does not appear to be implemented in the JavaScript or
Node.js integrations. I see a config option in the Node.js integration,
but it is unused, so I'm leaving out the guide for it.