Merge ignore actions pages
Merge the pages from the integrations that had separate ignore actions
pages.

I didn't add a specific feature page for "actions". We should do that in
the future, but right now that would be a lot of work and we can gain
more by focussing on restructuring the content we already have.

I couldn't find a page about this feature for the front-end JavaScript
integration, so I omitted it.