Add configuration guide
Add guide to help users configure AppSignal using an example config
option: debug/namespace.

This guide can be useful as first introduction to configuration in
AppSignal, but can also be linked from other parts of the docs and
guides to help users set any config options we mention in docs and
examples.